total_budget fixed

+++ cost_manager/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Budget(models.Model):
-    total_budget = models.DecimalField(max_digits=10, decimal_places=2)
+    total_budget = models.IntegerField()
 
     def __str__(self):
         return f"Total Budget: {self.total_budget}"
@@ -16,10 +16,9 @@ PURCHASE_STATUS = (
     (5, "امیرمحمد"),
 )
 
-
 class PurchaseItem(models.Model):
     name = models.CharField(max_length=200)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.IntegerField()
     quantity = models.IntegerField()
     purchase_link = models.URLField(null=True, blank=True)
     purchase_date = models.DateField()
